add auth checks for post CRUD and set post creator from requester

+++ src/main/java/com/team3/forum/services/PostServiceImpl.java
@@ -1,9 +1,11 @@
 package com.team3.forum.services;
 
+import com.team3.forum.exceptions.AuthorizationException;
 import com.team3.forum.exceptions.DuplicateEntityException;
 import com.team3.forum.exceptions.EntityNotFoundException;
 import com.team3.forum.models.Post;
 import com.team3.forum.models.User;
+import com.team3.forum.models.postDtos.PostUpdateDto;
 import com.team3.forum.repositories.PostRepository;
 import com.team3.forum.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +19,9 @@ import java.util.List;
 public class PostServiceImpl implements PostService{
     public static final String ALREADY_LIKED_ERROR = "This post is already liked by this user.";
     public static final String NOT_LIKED_ERROR = "This post is not liked by this user.";
-    
+    public static final String EDIT_AUTHORIZATION_ERROR = "You cannot edit this post.";
+    public static final String DELETE_AUTHORIZATION_ERROR = "You cannot delete this post.";
+
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
@@ -40,7 +44,11 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id, User requester) {
+        Post persistent = postRepository.findById(id);
+        if (!requester.isAdmin() && persistent.getUser().getId() != requester.getId()) {
+            throw new AuthorizationException(DELETE_AUTHORIZATION_ERROR);
+        }
         postRepository.deleteById(id);
     }
 
@@ -50,8 +58,15 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public Post update(Post post) {
-        return postRepository.save(post);
+    public Post update(int postId, PostUpdateDto postUpdateDto, User requester) {
+        Post persistent = postRepository.findById(postId);
+        if (!requester.isAdmin() && persistent.getUser().getId() != requester.getId()) {
+            throw new AuthorizationException(EDIT_AUTHORIZATION_ERROR);
+        }
+        persistent.setTitle(postUpdateDto.getTitle());
+        persistent.setContent(postUpdateDto.getContent());
+
+        return postRepository.save(persistent);
     }
 
     @Override
